added autorizations to endpoints

+++ src/main/java/blog_project/controller/MainController.java
@@ -9,6 +9,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -23,6 +24,7 @@ import org.springframework.web.servlet.ModelAndView;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -59,15 +61,14 @@ public class MainController {
 
     @GetMapping("/{id}")
     public String getRecipeById(@AuthenticationPrincipal  User user, @PathVariable Long id, Model model) {
-        boolean isAdmin = Optional.ofNullable(user).map(User::getRoles).map(roles -> roles.contains(Role.ADMIN)).orElse(false);
         Recipe recipe = recipeService.getRecipeById(id);
         model.addAttribute("recip", recipe);
-        model.addAttribute("isAdmin", isAdmin);
+        model.addAttribute("isAdmin", User.isAdmin(user));
         model.addAttribute("comment", new Comment());
         return "recipe";
     }
 
-
+    @PreAuthorize("hasAuthority('ADMIN')")
     @GetMapping("/form")
     public String getForm(@AuthenticationPrincipal  User user, Model model) {
         model.addAttribute("newRecipe", new Recipe());
@@ -75,7 +76,7 @@ public class MainController {
     }
 
 
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @PreAuthorize("hasAuthority('ADMIN')")
     @PostMapping("/form")
     public String processForm(@RequestParam("prdimage") MultipartFile file, @ModelAttribute Recipe newRecipe) throws IOException {
         String image = Base64.encodeBase64String(file.getBytes());
@@ -96,6 +97,7 @@ public class MainController {
         return "redirect:/mainPage/" + id;
     }
 
+    @PreAuthorize("hasAuthority('ADMIN')")
     @GetMapping("/recipes")
     public String getRecipes(Model model,
                              @PageableDefault(sort = { "title"}, direction = Sort.Direction.DESC, size = 2)
@@ -114,12 +116,14 @@ public class MainController {
         return "recipes";
     }
 
+    @PreAuthorize("hasAuthority('ADMIN')")
     @GetMapping("/recipes/delete/{id}")
     public String deleteRecipe(@PathVariable("id") long id) {
         recipeService.deleteRecipe(id);
         return "redirect:/mainPage/recipes";
     }
 
+    @PreAuthorize("hasAuthority('ADMIN')")
     @GetMapping("recipes/showUpdateForm")
     public ModelAndView showUpdateRecipeForm(@RequestParam Long id) {
         ModelAndView mav = new ModelAndView("edit_form");
@@ -128,6 +132,7 @@ public class MainController {
         return mav;
     }
 
+    @PreAuthorize("hasAuthority('ADMIN')")
     @PostMapping("/recipes/saveUpdatedRecipe")
     public String saveUpdatedRecipe(@RequestParam("prdimage") MultipartFile file, @ModelAttribute Recipe recipe) throws IOException {
         String image = Base64.encodeBase64String(file.getBytes());
@@ -137,26 +142,41 @@ public class MainController {
         return "redirect:/mainPage/recipes";
     }
 
+    @PreAuthorize("hasAnyAuthority('USER', 'ADMIN')")
     @GetMapping("/recipes/deleteComment/{id}")
-    public String deleteComment(@PathVariable("id") long id) {
+    public String deleteComment(@AuthenticationPrincipal User user, @PathVariable("id") long id) {
         Comment comment = commentService.findCommentById(id);
+        Long idRecipe = comment.getRecipe().getId();
+        if(user == null || (!comment.getCreatedBy().getId().equals(user.getId()) && !User.isAdmin(user))) {
+            throw new AccessDeniedException("Access denied");
+        }
         commentService.deleteCommentById(id);
-        return "redirect:/mainPage/"+ comment.getRecipe().getId();
+        return "redirect:/mainPage/"+ idRecipe;
     }
 
+    @PreAuthorize("hasAnyAuthority('USER', 'ADMIN')")
     @GetMapping("/recipes/commentEdit")
-    public ModelAndView showCommentForm(@RequestParam Long id) {
+    public ModelAndView showCommentForm(@AuthenticationPrincipal User user, @RequestParam Long id) {
         ModelAndView mav = new ModelAndView("recipe-comment-edit");
         Comment comment = commentService.findCommentById(id);
+        if(user == null || !comment.getCreatedBy().getId().equals(user.getId())) {
+            throw new AccessDeniedException("Access denied");
+        }
         mav.addObject("recipeee", comment);
         return mav;
     }
 
+    @PreAuthorize("hasAnyAuthority('USER', 'ADMIN')")
     @PostMapping("/recipes/saveComment")
-    public String saveComment(@ModelAttribute Comment comment) {
-        comment.setUpdatedDateTime(LocalDateTime.now());
-        commentService.addCommentToRecipe(comment);
-        return "redirect:/mainPage/" + comment.getRecipe().getId();
+    public String saveComment(@AuthenticationPrincipal User user, @ModelAttribute Comment comment) {
+        Comment savedComment = commentService.findCommentById(comment.getId());
+        if(user == null || !savedComment.getCreatedBy().getId().equals(user.getId())) {
+            throw new AccessDeniedException("Access denied");
+        }
+        savedComment.setText(comment.getText());
+        savedComment.setUpdatedDateTime(LocalDateTime.now());
+        commentService.addCommentToRecipe(savedComment);
+        return "redirect:/mainPage/" + savedComment.getRecipe().getId();
     }
 
     @GetMapping("/filterByTitle")
